cssImport false / pass select2 data in change event
cssImport changes the global context so setting it to true on one component in the app, loads the CSS for all. Disable it by default (I opt for removing it altogether).

Don't use `that` hack, use closures instead to get the desired context of `this`.

Pass the select2 data structure for each `valueChanged` event.