Bugfix: Multiselectvalues where not updated
When using this element with the "multiple" option, the value was not updated when it was changed from outside.

This was because "ngOnChanges" only performes a value checking not, a deep value checking. (Source: https://stackoverflow.com/questions/34796901/angular2-change-detection-ngonchanges-not-firing-for-nested-object ).
Beucause with the "multiple" option, "value" is a "string[]", a simple "ngOnChanges" check is not enough enough, it needs to check on "ngDoCheck".

Code:
    ngDoCheck() {
        if ($.isArray(this.value)) {
            if (this.oldValueLength !== this.value.length) {
                this.oldValueLength = this.value.length;
                this.setElementValue(this.value);
            }
        }
    }^


This code only check, if the value is an array and sets the new value in an array. To detect the change, 
    private oldValueLength = 0;
Is added, to keep track of the last lenght. (changes from null to array where already detected.)